SPEC simpoint collection
Due to some weird instruction errors, enough of the leela results are
missing to make them not very useful. There is a single x264 results
missing, I can't figure out why. The invocation looks right, and whilst
there is an error in the log, the same error appears in two other
checkpoint logs that produce output fine, just discard it and go ahead.